Abstract entity refactored
