+++ lib/index.js
@@ -1,11 +1,7 @@
 "use strict";
 
+var entity = require('./entity');
 
-function entity(root) {
-	return function (namespace, data) {
-		return null;
-	};
-}
 function is(root) {
 	return function (entity, namespace) {
 		return null;
