+++ package.json
@@ -15,8 +15,7 @@
 	],
 	"dependencies": {
 		"async": "~0.2.10",
-		"underscore": "~1.5.2",
-		"underscore.string": "~2.3.3"
+		"underscore": "~1.5.2"
 	},
 	"devDependencies": {
 		"grunt": "~0.4.2",
